moves everything back into the main project folder since flaskr is not actually a reserved name
